Make files correct
